extract pdp wishlist logic
